requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
httpx>=0.27
python-dotenv>=1.0
streamlit>=1.32
